cleanse date_strings before parsing to support precision and timezones #5
Resolves #5

Most times returned from OFX are not precise and typically include time at noon (12). For that reason, this choses to abandon precision and timezone knowledge to only report the posted _date_ and not time.
Most of my experience with the time returned is the same format with 14 digits. If more are discovered, we will add support for that as it comes.